requirements.txt
alembic==1.12.1
anyio==3.7.0
astroid==2.15.5
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.2
decorator==5.1.1
dill==0.3.6
distlib==0.3.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.95.2
filelock==3.12.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
identify==2.5.24
idna==3.4
iniconfig==2.0.0
isort==5.12.0
kafka-python==2.0.2
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
nodeenv==1.8.0
numpy==1.26.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.4
passlib==1.7.4
pbr==5.11.1
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
pylint==2.17.4
pytest==7.3.1
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
sqlalchemy-migrate==0.13.0
sqlalchemy2-stubs==0.0.2a34
sqlparse==0.4.4
starlette==0.27.0
Tempita==0.5.2
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.66.1
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.23.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
